requirements.txt
# Core dependencies
boto3>=1.26.0
botocore>=1.29.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# Optional dependencies for enhanced features
rich>=13.0.0

# AWS dependencies (optional for Comprehend Medical integration)
boto3>=1.26.0
botocore>=1.29.0

# Type checking and development
typing-extensions>=4.0.0
mypy>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.0.0 